requirements.txt
fastapi
uvicorn
joblib
pandas
numpy
scikit-learn
supabase
python-dotenv